move easy leaf nodes
