+++ pkg/image/apis/image/validation/whitelist/wildcard_test.go
@@ -1,4 +1,4 @@
-package strings
+package whitelist
 
 import "testing"
