+++ pkg/template/templateprocessing/object_test.go
@@ -1,4 +1,4 @@
-package stringreplace
+package templateprocessing
 
 import (
 	"fmt"
